requirements.txt
django==5.0.2
django-ckeditor==6.3.0
Pillow==10.2.0
